live_transcription/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0
aiofiles==23.2.1

# ML Core
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# SpeechBrain for diarization
speechbrain==1.0.0

# Audio utils
pydub==0.25.1
scipy==1.10.1

# Video + Docs
opencv-python==4.9.0.80
moviepy==1.0.3
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
pandas